monster shouts insults

+++ src/map.rs
@@ -21,7 +21,7 @@ pub struct Map {
 
 impl BaseMap for Map {
     fn is_opaque(&self, idx: usize) -> bool {
-        self.tiles[idx as usize] == TileType::Wall
+        self.tiles[idx] == TileType::Wall
     }
 }
